+++ kmm/src/commonMain/kotlin/fr/beapp/interviews/bicloo/kmm/logic/station/entity/PositionEntity.kt
@@ -1,0 +1,9 @@
+package fr.beapp.interviews.bicloo.kmm.logic.station.entity
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PositionEntity(
+	val latitude: Double,
+	val longitude: Double
+)