+++ kmm/src/commonMain/kotlin/fr/beapp/interviews/bicloo/kmm/logic/station/entity/StandEntity.kt
@@ -1,0 +1,9 @@
+package fr.beapp.interviews.bicloo.kmm.logic.station.entity
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class StandEntity(
+	val availabilities: AvailabilitiesEntity,
+	val capacity: Int
+)