+++ kmm/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `ContractEntity`, `ContractManager` and `ContractPresenter` classes
+- `StationEntity`,`StationManager` and `StationPresenter` classes
 
 ### Chore